@@ -12,7 +12,8 @@
     "start-prod-server": "cross-env NODE_ENV=production node server.js",
     "test": "nyc mocha",
     "coverage": "nyc report --reporter=text-lcov | coveralls",
-    "linter": "eslint --ext .jsx --ext .js -o lint-report.html src/"
+    "lint": "eslint src/**/*.js src/**/*.jsx",
+    "fix": "eslint --fix src/**/*.js src/**/*.jsx"
   },
   "author": "Mike Chabot",
   "license": "MIT",
@@ -35,66 +36,69 @@
     "report-dir": "./.nyc-coverage"
   },
   "dependencies": {
+    "@babel/polyfill": "^7.0.0",
     "axios": "^0.18.0",
-    "babel-polyfill": "^6.26.0",
-    "bulma": "^0.7.1",
+    "bulma": "^0.7.2",
     "cross-env": "^5.2.0",
-    "express": "^4.16.3",
+    "express": "^4.16.4",
     "glamor": "^2.20.40",
     "glamorous": "^4.13.1",
-    "lodash": "^4.17.10",
+    "lodash": "^4.17.11",
     "moment": "^2.22.2",
-    "normalize.css": "^8.0.0",
+    "normalize.css": "^8.0.1",
     "prop-types": "^15.6.2",
-    "react": "^16.4.1",
-    "react-dom": "^16.4.1",
-    "react-redux": "^5.0.7",
+    "react": "^16.6.3",
+    "react-dom": "^16.6.3",
+    "react-redux": "^5.1.1",
     "react-router-dom": "^4.3.1",
     "react-tabify": "0.1.24",
-    "redux": "^4.0.0",
+    "redux": "^4.0.1",
     "redux-entity": "^5.1.2",
     "redux-logger": "^3.0.6",
     "redux-thunk": "^2.3.0"
   },
   "devDependencies": {
-    "autoprefixer": "^8.6.5",
-    "babel-core": "^6.26.3",
-    "babel-loader": "^7.1.5",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-0": "^6.24.1",
-    "chai": "^4.1.2",
+    "@babel/cli": "^7.1.5",
+    "@babel/core": "^7.1.6",
+    "@babel/preset-env": "^7.1.6",
+    "@babel/preset-react": "^7.0.0",
+    "@babel/register": "^7.0.0",
+    "autoprefixer": "^9.3.1",
+    "babel-eslint": "^10.0.1",
+    "babel-loader": "^8.0.4",
+    "babel-minify-webpack-plugin": "^0.3.1",
+    "chai": "^4.2.0",
     "coveralls": "^3.0.2",
-    "css-loader": "^1.0.0",
+    "css-loader": "^1.0.1",
     "del": "^3.0.0",
-    "eslint": "^5.1.0",
-    "eslint-config-standard": "^12.0.0-alpha.0",
-    "eslint-loader": "^2.0.0",
-    "eslint-plugin-import": "^2.13.0",
-    "eslint-plugin-node": "^6.0.1",
-    "eslint-plugin-promise": "^3.8.0",
-    "eslint-plugin-react": "^7.10.0",
-    "eslint-plugin-standard": "^3.1.0",
-    "file-loader": "^1.1.11",
+    "eslint": "^5.9.0",
+    "eslint-config-standard": "^12.0.0",
+    "eslint-loader": "^2.1.1",
+    "eslint-plugin-import": "^2.14.0",
+    "eslint-plugin-node": "^8.0.0",
+    "eslint-plugin-promise": "^4.0.1",
+    "eslint-plugin-react": "^7.11.1",
+    "eslint-plugin-standard": "^4.0.0",
+    "file-loader": "^2.0.0",
     "html-webpack-plugin": "^3.2.0",
-    "jsdom": "^11.11.0",
+    "jsdom": "^13.0.0",
     "json-loader": "^0.5.7",
-    "mini-css-extract-plugin": "^0.4.1",
+    "mini-css-extract-plugin": "^0.4.4",
     "mocha": "^5.2.0",
-    "node-sass": "^4.9.2",
-    "nyc": "^12.0.2",
-    "postcss-loader": "^2.1.6",
+    "node-sass": "^4.10.0",
+    "nyc": "^13.1.0",
+    "postcss-loader": "^3.0.0",
     "react-addons-test-utils": "^15.6.2",
-    "react-hot-loader": "^4.3.3",
+    "react-hot-loader": "^4.3.12",
     "redux-mock-store": "^1.5.3",
-    "sass-loader": "^7.0.3",
-    "sinon": "^6.1.3",
-    "style-loader": "^0.21.0",
-    "svg-react-loader": "^0.4.5",
-    "uglifyjs-webpack-plugin": "^1.2.7",
-    "url-loader": "^1.0.1",
-    "webpack": "^4.16.0",
-    "webpack-cli": "^3.0.8",
-    "webpack-dev-server": "^3.1.4"
+    "sass-loader": "^7.1.0",
+    "sinon": "^7.1.1",
+    "style-loader": "^0.23.1",
+    "svg-react-loader": "^0.4.6",
+    "uglifyjs-webpack-plugin": "^2.0.1",
+    "url-loader": "^1.1.2",
+    "webpack": "^4.25.1",
+    "webpack-cli": "^3.1.2",
+    "webpack-dev-server": "^3.1.10"
   }
 }
