@@ -1,5 +1,5 @@
 module.exports = {
-    __hasValue: function (val) {
-        return val !== undefined && val !== null;
-    }
+  __hasValue: function (val) {
+    return val !== undefined && val !== null;
+  }
 };
