@@ -2,5 +2,5 @@ import React from 'react';
 import Icon from './Icon';
 
 export default function LoadingIndicator () {
-    return <Icon icon="cog fa-spin fa-2x" prefix="fas" />;
+  return <Icon icon="cog fa-spin fa-2x" prefix="fas"/>;
 }
