@@ -1,9 +1,9 @@
 import React from 'react';
 
 export default function NotFound () {
-    return (
-        <section>
-            <h5>404 Not Found</h5>
-        </section>
-    );
+  return (
+    <section>
+      <h5>404 Not Found</h5>
+    </section>
+  );
 }
