@@ -4,34 +4,34 @@ import RouteContent from './RouteContent';
 import { Tab, Tabs } from 'react-tabify';
 import { Route } from 'react-router-dom';
 
-function TabbedRouter ({
-    location,
-    history
-}) {
-    return (
-        <div className="notification m-top--small">
-            <Tabs
-                id="router-example-tabs"
-                activeKey={location.pathname}
-                onSelect={(eventKey) => history.push(eventKey)}>
-                <Tab eventKey="/" label="Increment">
-                    <Route exact path="/" component={RouteContent}/>
-                </Tab>
-                <Tab eventKey="/decrement" label="Decrement">
-                    <Route path="/decrement" component={RouteContent}/>
-                </Tab>
-                <Tab eventKey="/reset" label="Reset">
-                    <Route path="/reset" component={RouteContent}/>
-                </Tab>
-            </Tabs>
-        </div>
+const TabbedRouter = ({
+  location,
+  history
+}) => {
+  return (
+    <div className="notification m-top--small">
+      <Tabs
+        id="router-example-tabs"
+        activeKey={location.pathname}
+        onSelect={(eventKey) => history.push(eventKey)}>
+        <Tab eventKey="/" label="Increment">
+          <Route exact path="/" component={RouteContent}/>
+        </Tab>
+        <Tab eventKey="/decrement" label="Decrement">
+          <Route path="/decrement" component={RouteContent}/>
+        </Tab>
+        <Tab eventKey="/reset" label="Reset">
+          <Route path="/reset" component={RouteContent}/>
+        </Tab>
+      </Tabs>
+    </div>
 
-    );
-}
+  );
+};
 
 TabbedRouter.propTypes = {
-    location: PropTypes.object.isRequired,
-    history : PropTypes.object.isRequired
+  location: PropTypes.object.isRequired,
+  history: PropTypes.object.isRequired
 };
 
 export default TabbedRouter;
