@@ -1,7 +1,7 @@
 import {
-    INCREMENT_COUNTER,
-    DECREMENT_COUNTER,
-    RESET_COUNTER
+  INCREMENT_COUNTER,
+  DECREMENT_COUNTER,
+  RESET_COUNTER
 } from './types';
 
 export const increment = makeActionCreator(INCREMENT_COUNTER);
@@ -18,12 +18,12 @@ export const reset = makeActionCreator(RESET_COUNTER);
  *            actionCreator(123); --> { type: "DO_IT", data: 123 }
  */
 export function makeActionCreator (type, ...keys) {
-    if (!type) throw new Error('Type cannot be null/undefined');
-    return function (...args) {
-        let action = { type };
-        keys.forEach((arg, index) => {
-            action[keys[index]] = args[index];
-        });
-        return action;
-    };
+  if (!type) throw new Error('Type cannot be null/undefined');
+  return function (...args) {
+    let action = { type };
+    keys.forEach((arg, index) => {
+      action[keys[index]] = args[index];
+    });
+    return action;
+  };
 }
