@@ -3,6 +3,6 @@ import { entities } from 'redux-entity';
 import counter from './counter';
 
 export default combineReducers({
-    entities,
-    counter
+  entities,
+  counter
 });
