@@ -1,7 +1,7 @@
-import { expect } from 'chai';
+import {expect} from 'chai';
 
 describe('Integration Test', () => {
-    it('should pass the dummy test', () => {
-        expect(true).to.equal(true);
-    });
+  it('should pass the dummy test', () => {
+    expect(true).to.equal(true);
+  });
 });
